I made edge detection functions better
The first method finds the shape with the largest area and selects it.

The second method scores criteria such as proximity to the center, circularity, size, and finds which edge detection method to choose based on that score.

But still has problems. Especially on triangular shapes. I still don't know what to do.

After yahya murat joins, I'll switch branch